deliv-3 separate styled-components from component logic + more fixes
