+++ src/components/AvatarLogo/AvatarLogo.component.jsx
@@ -1,22 +1,8 @@
-import React, { useContext, useState, useEffect } from 'react';
-import styled from 'styled-components';
-
-import { AppearanceContext } from '../../contexts/AppearanceContextProvider';
+import React, { useContext } from 'react';
+import Styled from './AvatarLogo.styled';
 import { AccountContext } from '../../contexts/AccountContextProvider';
 
-const AvatarImg = styled.img`
-  display: inline-block;
-  max-width: 100%;
-  max-height: 100%;
-  /* background: url(${(props) => props.src}); */
-  /* background-image: url('../../../public/empty_avatar.png');
-  background-position: 50% 50%;
-  background-size: cover; */
-  border-radius: 50%;
-`;
-
-function AvatarLogo(props) {
-  const darkModeContext = useContext(AppearanceContext);
+function AvatarLogo() {
   const loggedInContext = useContext(AccountContext);
 
   const img = loggedInContext.account
@@ -25,10 +11,9 @@ function AvatarLogo(props) {
 
   return (
     <span>
-      <AvatarImg src={img} />
+      <Styled.AvatarImg src={img} />
     </span>
   );
-  // return <img src="../../../public/empty_avatar.png" alt="" />;
 }
 
 export default AvatarLogo;
