+++ src/components/AvatarLogo/AvatarLogo.test.js
@@ -1,8 +1,7 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import { BrowserRouter as Router } from 'react-router-dom';
-import CardItem from './index';
-// import mockData from './test_data/mock.json';
+// import React from 'react';
+// import { render, screen } from '@testing-library/react';
+// import { BrowserRouter as Router } from 'react-router-dom';
+// // import mockData from './test_data/mock.json';
 
-describe('AvatarLogo Component', () => {
-});
+// // describe('AvatarLogo Component', () => {
+// // });
