--- src/components/CardTitleText/index.js
@@ -1,1 +1,0 @@
-export { default } from './CardTitleText.component';
