+++ src/components/CenteredImage/CenteredImage.component.jsx
@@ -1,16 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-
-const CenteredImageWrapper = styled.img`
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: fit-content;
-  height: fit-content;
-`;
+import Styled from './CenteredImage.styled';
 
 function CenteredImage(props) {
-  return <CenteredImageWrapper src={props.image} />;
+  return <Styled.CenteredImageWrapper src={props.image} />;
 }
 
 export default CenteredImage;
