+++ src/components/CenteredImage/CenteredImage.styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const CenteredImageWrapper = styled.img`
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  width: fit-content;
+  height: fit-content;
+`;
+
+const Styled = {
+  CenteredImageWrapper,
+};
+
+export default Styled;
