--- src/components/Fortune/Fortune.styles.css
@@ -1,5 +1,0 @@
-.fortune {
-  position: fixed;
-  bottom: 0;
-  right: 0;
-}
