--- src/components/Fortune/index.js
@@ -1,1 +1,0 @@
-export { default } from './Fortune.component';
