--- src/components/ImageHover/index.js
@@ -1,1 +1,0 @@
-export { default } from './ImageHover.component';
