+++ src/components/Layout/Layout.styled.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+const LayoutWrapper = styled.main`
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 0rem;
+`;
+
+const Styled = {
+  LayoutWrapper,
+};
+
+export default Styled;
