+++ src/components/LoadingSpinner/LoadingSpinner.styled.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const Spinner = styled.img`
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  width: fit-content;
+  height: fit-content;
+`;
+
+const Styled = { Spinner };
+
+export default Styled;
