+++ src/components/LoginModal/LoginModal.component.jsx
@@ -1,61 +1,9 @@
 import React, { useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-import styled from 'styled-components';
+import Styled from './LoginModal.styled';
 import loginApi from '../../login.api';
-
 import { AccountContext } from '../../contexts/AccountContextProvider';
 import { storage } from '../../utils/storage';
-
-const ModalBackground = styled.div`
-  display: block;
-  position: fixed;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgb(0,0,0);
-  background-color: rgba(0,0,0,0.4);
-`;
-
-const XButton = styled.span`
-  color: #aaa;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-
-  &:hover,
-  &:focus {
-    color: black;
-    text-decoration: none;
-    cursor: pointer;
-  }
-`;
-
-const Input = styled.input`
-  margin: 10px 0;
-  border-radius: 5px;
-  border: 1px solid black;
-  line-height: 1.5em;
-  padding: 5px;
-`;
-
-const LoginContainer = styled.div`
-  height: 40%;
-  width: 40%;
-  float: right;
-  margin-right: 50%;
-  margin-left: 50%;
-  margin-top: 10%;
-  align-items: center;
-  background-color: ${(props) =>
-    props.theme.darkMode ? `rgb(0, 52, 85)` : `rgb(29, 106, 154, 1)`};
-`;
-
-const ErrorMessage = styled.p`
-  color: red;
-`;
 
 function LoginModal(props) {
   const loginContext = useContext(AccountContext);
@@ -88,20 +36,20 @@ function LoginModal(props) {
   };
 
   return (
-    <ModalBackground data-testid="login-modal-background">
-      <LoginContainer>
-        <XButton onClick={goBack}>&times;</XButton>
+    <Styled.ModalBackground data-testid="login-modal-background">
+      <Styled.LoginContainer>
+        <Styled.XButton onClick={goBack}>&times;</Styled.XButton>
         <h2>Welcome!</h2>
-        <ErrorMessage>{hasError}</ErrorMessage>
+        <Styled.ErrorMessage>{hasError}</Styled.ErrorMessage>
         <form onSubmit={handleLogin}>
           <h3>Username</h3>
-          <Input
+          <Styled.Input
             type="text"
             placeholder="username"
             onChange={(e) => setUser(e.target.value)}
           />
           <h3>Password</h3>
-          <Input
+          <Styled.Input
             type="password"
             placeholder="password"
             onChange={(e) => setPassword(e.target.value)}
@@ -110,8 +58,8 @@ function LoginModal(props) {
             Login
           </button>
         </form>
-      </LoginContainer>
-    </ModalBackground>
+      </Styled.LoginContainer>
+    </Styled.ModalBackground>
   );
 }
