+++ src/components/RelatedVideos/RelatedVideos.styled.js
@@ -1,0 +1,26 @@
+import styled from 'styled-components';
+
+const RelatedVideosWrapper = styled.div`
+  height: fit-content;
+  text-align: center;
+  width: 100%;
+  float: right;
+  margin: 0px;
+  margin-left: auto;
+  margin-right: auto;
+  float: left;
+  display: flex;
+  flex-direction: column;
+`;
+
+const EmptyDiv = styled.div`
+  width: 360px;
+  height: 112px;
+`;
+
+const Styled = {
+  RelatedVideosWrapper,
+  EmptyDiv,
+};
+
+export default Styled;
