+++ src/components/VideoEmbed/VideoEmbed.styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const VideoIframe = styled.iframe`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+`;
+
+const Styled = {
+  VideoIframe,
+};
+
+export default Styled;
