+++ src/components/VideoLayout/VideoLayout.component.jsx
@@ -1,20 +1,8 @@
 import React from 'react';
-import styled from 'styled-components';
-
-const LayoutWrapper = styled.main`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-top: 0rem;
-  margin-right: 0px;
-  overflow: auto;
-`;
+import Styled from './VideoLayout.styled';
 
 function VideoLayout({ children }) {
-  return <LayoutWrapper>{children}</LayoutWrapper>;
+  return <Styled.LayoutWrapper>{children}</Styled.LayoutWrapper>;
 }
 
 export default VideoLayout;
